Move args to actions
