+++ clcache/clcache_lib/clcache_lib.egg-info/PKG-INFO
@@ -1,6 +1,6 @@
 Metadata-Version: 2.1
 Name: clcache-lib
-Version: 4.4.4
+Version: 4.4.5
 Summary: ClCache Library
 Home-page: https://github.com/dgehri/clcache
 Author: Various
