+++ clcache/clcache_lib/clcache_lib/config/config.py
@@ -1,7 +1,7 @@
 from datetime import timedelta
 import os
 
-VERSION = "4.4.4-dgehri"
+VERSION = "4.4.5-dgehri"
 CACHE_VERSION = "9"
 
 COUCHBASE_EXPIRATION = timedelta(days=3)
