+++ clcache_launcher/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "clcache_launcher"
-version = "4.4.4"
+version = "4.4.5"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
